pyproject.toml
[tool.poetry]
name = "favbook"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-filter = "^23.2"
python-slugify = "^8.0.1"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
django-cleanup = "^8.0.0"
pytest-django = "^4.5.2"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
